requirements.txt
tensorflow
keras
scipy
requests
pycrypto
pillow
